# partita-service/src/main/java/com/connectedgames/core/dto/PartitaSyncInput.java
package com.connectedgames.core.dto;

import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import java.time.OffsetDateTime;
import java.util.UUID;

public record PartitaSyncInput(
    @NotNull UUID id,
    @NotBlank String installazioneId,
    @NotBlank String localeId,
    UUID giocatore1Id,
    String giocatore1Username,
    UUID giocatore2Id,
    String giocatore2Username,
    @NotNull @Min(0) @Max(999) Integer punteggio1,
    @NotNull @Min(0) @Max(999) Integer punteggio2,
    @NotNull OffsetDateTime dataInizio,
    @NotNull OffsetDateTime dataFine,
    UUID torneoId
) {}