# service-gateway/src/main/resources/static/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Connected Games Platform - Central Server API
  description: |
    Specifica ufficiale delle API REST del Server Centrale per il progetto Connected Games Platform (PISSIR A.A. 2025/2026).
    Il Server Centrale (dietro al Gateway) riceve le richieste dai Gateway Edge e dai client.
    Le rotte sono protette tramite token JWT emessi da Keycloak.
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/v1
    description: Server Gateway API (Sviluppo Locale)
paths:

  /locali/{localeId}/partite/sincronizza:
    post:
      summary: Sincronizzazione massiva offline (Bulk Sync)
      description: |
        Riceve un array di partite salvate offline nel buffer SQLite di uno specifico Edge.
        Il Server Centrale processa le partite singolarmente. Implementa una logica di tolleranza all'errore
        restituendo un successo parziale (Bulk Partial Success) per evitare loop infiniti di sincronizzazione.
        L'idempotenza è garantita dall'UUID di ciascuna partita generato a monte dall'Edge.
      security:
        - BearerAuth: []
      parameters:
        - name: localeId
          in: path
          required: true
          description: ID del locale fisico che sta inviando i dati (es. BAR_BELVEDERE)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PartitaSyncInput'
      responses:
        '200':
          description: Sincronizzazione elaborata (può contenere successi e fallimenti parziali)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResultResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '500':
          $ref: '#/components/responses/500InternalError'

  /locali/{localeId}/giochi:
    get:
      summary: Lista dei giochi installati nel locale
      description: Recupera tutti i giochi fisici censiti e attivi all'interno di uno specifico locale per la sincronizzazione o la visualizzazione di stato.
      security:
        - BearerAuth: []
      parameters:
        - name: localeId
          in: path
          required: true
          description: ID univoco del locale fisico
          schema:
            type: string
      responses:
        '200':
          description: Lista dei giochi recuperata con successo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GiocoInstallato'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          description: Locale non trovato
        '500':
          $ref: '#/components/responses/500InternalError'

  /tornei:
    get:
      summary: Ottieni la lista dei tornei
      description: Restituisce l'elenco di tutti i tornei registrati sulla piattaforma, attivi o conclusi.
      security:
        - BearerAuth: []
      parameters:
        - name: stato
          in: query
          required: false
          description: Filtra i tornei per stato (es. ATTIVO, CONCLUSO)
          schema:
            type: string
            enum: [ATTIVO, CONCLUSO]
      responses:
        '200':
          description: Lista dei tornei recuperata con successo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Torneo'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalError'
    post:
      summary: Crea un nuovo torneo
      description: Permette la creazione di un torneo associato a un gioco e a vari locali (admin only).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TorneoCreateInput'
      responses:
        '200':
          description: Torneo creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Torneo'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '500':
          $ref: '#/components/responses/500InternalError'

  /tornei/{torneoId}:
    get:
      summary: Recupera i dettagli di un singolo torneo
      description: Restituisce le informazioni di un torneo specifico.
      security:
        - BearerAuth: []
      parameters:
        - name: torneoId
          in: path
          required: true
          description: ID univoco del torneo
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dettagli torneo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Torneo'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          description: Torneo non trovato
        '500':
          $ref: '#/components/responses/500InternalError'

  /tornei/{torneoId}/iscrizioni:
    get:
      summary: Ottieni iscritti al torneo
      description: Restituisce la lista degli iscritti al torneo specificato.
      security:
        - BearerAuth: []
      parameters:
        - name: torneoId
          in: path
          required: true
          description: ID univoco del torneo
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista iscritti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IscrizioneTorneoResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalError'
    post:
      summary: Iscrivi un giocatore al torneo
      description: Permette l'iscrizione di un giocatore a un torneo attivo.
      security:
        - BearerAuth: []
      parameters:
        - name: torneoId
          in: path
          required: true
          description: ID univoco del torneo
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IscrizioneInput'
      responses:
        '200':
          description: Giocatore iscritto con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IscrizioneTorneoResponse'
        '400':
          description: Dati non validi o giocatore già iscritto o torneo chiuso
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalError'

  /tornei/{torneoId}/classifica:
    get:
      summary: Ottieni la classifica in tempo reale del torneo
      description: |
        Calcola live (lazy evaluation) la classifica aggregata del torneo sommando o normalizzando i punteggi
        delle partite validate giocate nei diversi locali coinvolti.
        Non richiede materializzazione di tabelle.
      security:
        - BearerAuth: []
      parameters:
        - name: torneoId
          in: path
          required: true
          description: ID univoco del torneo
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Classifica del torneo calcolata in tempo reale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificaTorneoResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          description: Torneo non trovato
        '500':
          $ref: '#/components/responses/500InternalError'

  /statistiche:
    get:
      summary: Dashboard delle statistiche globali della piattaforma
      description: |
        Restituisce un report aggregato calcolato in tempo reale (live queries) per gli Amministratori di Piattaforma.
        Include metriche di utilizzo complessivo, classifiche globali dei giocatori e popolarità dei giochi.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistiche globali aggregate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticheGlobaliResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          description: Accesso negato. Solo gli Amministratori di Piattaforma hanno i permessi per visualizzare le statistiche globali.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/500InternalError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Inserisci il token JWT rilasciato da Keycloak.

  responses:
    401Unauthorized:
      description: Token mancante, scaduto o non valido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    403Forbidden:
      description: Privilegi insufficienti per accedere alla risorsa.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    500InternalError:
      description: Errore interno del server durante l'elaborazione.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-07-12T09:17:00Z"
        status:
          type: integer
          example: 401
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "Full authentication is required to access this resource"

    PartitaSyncInput:
      type: object
      required:
        - id
        - giocoId
        - punteggio1
        - punteggio2
        - dataInizio
        - dataFine
      properties:
        id:
          type: string
          format: uuid
          description: UUID univoco della partita generato a monte dall'Edge (fondamentale per garantire l'idempotenza).
          example: "7e504c86-1d18-4e89-9a73-31f1dcce7780"
        giocoId:
          type: string
          description: Identificativo del gioco installato nel locale (es. calciobalilla-verde).
          example: "calciobalilla-1"
        giocatore1Id:
          type: string
          format: uuid
          nullable: true
          description: ID dell'utente Keycloak per il giocatore 1. Se NULL, identifica un giocatore Ospite (Guest Mode).
          example: "a3b9cc22-38e2-4bf1-a06f-b1b7ef38f123"
        giocatore2Id:
          type: string
          format: uuid
          nullable: true
          description: ID dell'utente Keycloak per il giocatore 2. Se NULL, identifica un giocatore Ospite (Guest Mode).
          example: "e883df11-20f5-4c07-ba99-8cfc84381e44"
        punteggio1:
          type: integer
          description: Punteggio finale ottenuto dal giocatore 1 o dal Team A.
          example: 10
        punteggio2:
          type: integer
          description: Punteggio finale ottenuto dal giocatore 2 o dal Team B.
          example: 8
        dataInizio:
          type: string
          format: date-time
          description: Timestamp ISO-8601 UTC dell'inizio effettivo della partita.
          example: "2026-07-12T09:00:00Z"
        dataFine:
          type: string
          format: date-time
          description: Timestamp ISO-8601 UTC della conclusione della partita.
          example: "2026-07-12T09:15:00Z"
        torneoId:
          type: string
          format: uuid
          nullable: true
          description: Valorizzato solo se la partita faceva parte di un torneo ufficiale attivo al momento del gioco.
          example: "303fde9b-0268-4682-8302-39bd123f1234"

    SyncResultResponse:
      type: object
      required:
        - salvate
        - fallite
      properties:
        salvate:
          type: array
          description: Array contenente gli UUID delle partite sincronizzate e salvate con successo in PostgreSQL. L'Edge userà questa lista per fare la DELETE immediata dal suo SQLite locale.
          items:
            type: string
            format: uuid
          example: ["7e504c86-1d18-4e89-9a73-31f1dcce7780"]
        fallite:
          type: array
          description: Array delle partite che hanno riscontrato errori durante l'elaborazione (es. violazione di regole di scoring o incoerenza dati). L'Edge le escluderà dai prossimi invii per evitare blocchi della coda.
          items:
            $ref: '#/components/schemas/SyncFailure'
          example:
            - id: "bc34fe81-e009-41aa-8bc3-32f41ccfe999"
              errore: "Il torneo_id indicato è scaduto o inesistente"

    SyncFailure:
      type: object
      required:
        - id
        - errore
      properties:
        id:
          type: string
          format: uuid
          example: "bc34fe81-e009-41aa-8bc3-32f41ccfe999"
        errore:
          type: string
          example: "Il punteggio inserito non rispetta le regole del gioco Calciobalilla (punteggio massimo 10)"

    GiocoInstallato:
      type: object
      required:
        - id
        - tipoGioco
      properties:
        id:
          type: string
          example: "calciobalilla-1"
        tipoGioco:
          type: string
          enum: [CALCIOBALILLA, FRECCETTE]
          example: "CALCIOBALILLA"

    Torneo:
      type: object
      required:
        - id
        - nome
        - giocoNome
        - stato
        - dataInizio
        - dataFine
      properties:
        id:
          type: string
          format: uuid
          example: "303fde9b-0268-4682-8302-39bd123f1234"
        nome:
          type: string
          example: "Torneo Estivo UPO Calciobalilla 2026"
        giocoNome:
          type: string
          example: "CALCIOBALILLA"
        stato:
          type: string
          enum: [ATTIVO, CONCLUSO]
          example: "ATTIVO"
        dataInizio:
          type: string
          format: date-time
          example: "2026-07-01T00:00:00Z"
        dataFine:
          type: string
          format: date-time
          example: "2026-07-15T23:59:59Z"

    TorneoCreateInput:
      type: object
      required:
        - nome
        - giocoId
        - localiId
        - dataInizio
        - dataFine
      properties:
        nome:
          type: string
          example: "Campionato Nazionale"
        giocoId:
          type: string
          example: "calciobalilla"
        localiId:
          type: array
          items:
            type: string
          example: ["BAR_BELVEDERE"]
        dataInizio:
          type: string
          format: date-time
        dataFine:
          type: string
          format: date-time

    IscrizioneInput:
      type: object
      required:
        - utenteId
      properties:
        utenteId:
          type: string
          format: uuid
          example: "a3b9cc22-38e2-4bf1-a06f-b1b7ef38f123"

    IscrizioneTorneoResponse:
      type: object
      required:
        - torneoId
        - utenteId
        - utenteNome
        - dataIscrizione
      properties:
        torneoId:
          type: string
          format: uuid
        utenteId:
          type: string
          format: uuid
        utenteNome:
          type: string
          example: "mario_rossi"
        dataIscrizione:
          type: string
          format: date-time

    ClassificaTorneoResponse:
      type: object
      required:
        - torneoId
        - nomeTorneo
        - tipoGioco
        - classifica
      properties:
        torneoId:
          type: string
          format: uuid
          example: "303fde9b-0268-4682-8302-39bd123f1234"
        nomeTorneo:
          type: string
          example: "Torneo Estivo UPO Calciobalilla 2026"
        tipoGioco:
          type: string
          example: "CALCIOBALILLA"
        classifica:
          type: array
          items:
            $ref: '#/components/schemas/RigaClassifica'

    RigaClassifica:
      type: object
      required:
        - posizione
        - utenteId
        - username
        - partiteGiocate
        - metricaClassifica
      properties:
        posizione:
          type: integer
          example: 1
        utenteId:
          type: string
          format: uuid
          example: "a3b9cc22-38e2-4bf1-a06f-b1b7ef38f123"
        username:
          type: string
          example: "mario_rossi"
        partiteGiocate:
          type: integer
          example: 12
        metricaClassifica:
          type: number
          format: float
          description: |
            Rappresenta la metrica normalizzata per la classifica.
            Per il calciobalilla indica la percentuale di vittorie (o media gol/partita) per evitare vantaggi a chi gioca più partite.
            Per le freccette indica il punteggio cumulato.
          example: 0.83

    StatisticheGlobaliResponse:
      type: object
      required:
        - totalePartiteGiocate
        - totaleGiocatoriAttivi
        - localiPiuAttivi
        - giochiPiuUtilizzati
      properties:
        totalePartiteGiocate:
          type: integer
          example: 1420
        totaleGiocatoriAttivi:
          type: integer
          example: 256
        localiPiuAttivi:
          type: array
          items:
            type: object
            properties:
              localeId:
                type: string
                example: "BAR_BELVEDERE"
              nomeLocale:
                type: string
                example: "Bar Belvedere"
              partiteGiocate:
                type: integer
                example: 612
        giochiPiuUtilizzati:
          type: array
          items:
            type: object
            properties:
              giocoTipo:
                type: string
                example: "CALCIOBALILLA"
              partiteGiocate:
                type: integer
                example: 980